requirements.txt
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
